Update getMe tests
